Nav bar and intro heading
Create a responsive navbar.

Navbar collapse and become a hamburger.

Navbar items are align to the left.

Add dogo as website logo.

Increase font size for the title in index.html Each letter in this title is made of span tag and each letter will be style in different color later.